Performance Tuning
This push adds the InitialMeasurementTest, which was used to determine initial compute time. ComputeControllerAPIFast, an alternate version of ComputeControllerAPI that caches results, so that if they are repeated they're pulled from cache instead of being computed repeatedly. ComputeEngineFastTest, which tests performance of old in comparison to the fast API and displays the time difference, as well as an assertion to confirm an at least 10% improvement.