Memory fix (#52)
* Performance Tuning

This push adds the InitialMeasurementTest, which was used to determine initial compute time. ComputeControllerAPIFast, an alternate version of ComputeControllerAPI that caches results, so that if they are repeated they're pulled from cache instead of being computed repeatedly. ComputeEngineFastTest, which tests performance of old in comparison to the fast API and displays the time difference, as well as an assertion to confirm an at least 10% improvement.

* Update ComputeControllerAPIFast.java

Fixing error that set off checkpoint 3 test.

* Update ComputeFastTest.java

* Fixing checkpoint 3 error problem

Because of the implementation of API, i believe this confused the checkpoint 3 test, because there should only be one so i changed the structure, so that would not be the case.

* Update README.md

Adding performance tuning section detailing bottleneck and fix.

* Update README with performance test information

Added performance test reference to README.

* Rename ComputeFastTest to ComputeEngineFastTest

* Update link to performance test file in README

* Update README with hardware and performance info

Added hardware specifications and performance improvement details.

* Memory Fix

Fixing the client's ability to use memory as input type successfully, as well as some other minor tweaks.

* Fixing test issues

fixing test issues by restoring file structure.

* Update DataStoreComputeAPIImplementation.java

* Performance Tuning update

Updating cache so that it has a bounded size and an actual eviction policy in place. Also updating test with a mix of inputs for more effective testing, instead of same n.

* Resolve issue

* Update ComputeEngineFastTest.java

* Update ComputeEngineFastTest.java

* Update ComputeEngineFastTest.java

* Update ComputeEngineFastTest.java