EndToEndGrpcTest(final)
This test starts datastore gRPC server with the datastore implementation. It then connects a channel to it and wraps it with the GrpcDataStoreComputeAPI. A compute gRPC server with the computecontrollerapiimplementation and the user compute implementation. Build a UserSubmission just like client. Calls .submit over gRPC. Assertions to ensure pipline is working as intended. This accomplishes thorough testing over the gRPC from end to end, verifying behavior works as intended.